backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
Werkzeug==3.1.3
python-dotenv==1.0.1

# Production server
gunicorn==21.2.0

# Optional database drivers (install separately if needed)
# For PostgreSQL: uv pip install psycopg2-binary
# For MySQL: uv pip install PyMySQL
# SQLite is included with Python by default
